+++ src/environments/deeprl/simpledeepmarlenv.py
@@ -190,7 +190,7 @@ class SimplePersuasiveDeepMARLEnv(PersuasiveDeepMARLEnv):
         # Flattening of the dictionary
         deep_ret = self.deep_state_flattener(ret)
         logger.debug('[%s] Observation: %s', agent, str(deep_ret))
-        return np.array(deep_ret, dtype=np.int64)
+        return np.array(deep_ret, dtype=np.float64)
 
     def get_obs_space(self, agent):
         """ Returns the observation space. """
